fix: remove `?` from plugin name in docs (#2)
The name attribute in docs is used by woodpecker docs as the slug for the plugin documentation. This means a slug with ? is generated for the plugin, but the ? is interpreted as the start of query params in the URL and not part of slug, which means the plugin page is not accessible.